requirements.txt
# Dependencias para ETL MySQL Project
# Instalar con: pip install -r requirements.txt

# Conexión a MySQL
mysql-connector-python==8.2.0

# Manipulación de datos
pandas==2.2.1
numpy==1.26.4

# Generación de datos sintéticos
faker==22.0.0

# Variables de entorno (opcional pero recomendado)
python-dotenv==1.0.0

# Logging mejorado
loguru==0.7.2
